% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WQSdata}
\alias{WQSdata}
\title{Simulated data of chemical concentrations and one continuous outcome variable}
\format{A data frame with 1000 rows and 10 variables:
\describe{
\item{X1}{a numeric vector}
\item{X2}{a numeric vector}
\item{X3}{a numeric vector}
\item{X4}{a numeric vector}
\item{X5}{a numeric vector}
\item{X6}{a numeric vector}
\item{X7}{a numeric vector}
\item{X8}{a numeric vector}
\item{X9}{a numeric vector}
\item{y}{a numeric vector; the outcome variable}
}}
\usage{
WQSdata
}
\description{
Correlation and concentration patterns were loosely based on NHL data.
}
\keyword{datasets}
